pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "facegazesynth"
version = "0.1.0"
description = "Physics-based synthetic face and eye gaze renderer"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "matplotlib>=3.7",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["facegazesynth*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
